UART: refactor AVR and NRF implementations to use shared code implementation
